(송인석) 컨트롤러 보안작업
BookController.java - 세션값 확인을 통해 보안작업
CityController.java - 세션값 확인을 통해 보안작업
CountryController.java - 세션값 확인을 통해 보안작업
IdolController.java - 디버그 확인하는 부분 수정

+++ test26/src/main/java/ksmart/project/test26/BookController.java
@@ -2,6 +2,8 @@ package ksmart.project.test26;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -22,7 +24,11 @@ public class BookController {
 	private BookDao bookDao;
 	
 	@RequestMapping(value="/book/bookList")
-	public String BookList(Model model) {
+	public String BookList(Model model,HttpSession session) {
+		// 세션에 로그인 값을 확인하고 로그인 정보가 없으면 리다이렉트
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect://login1team/login";
+		}
 		List<Book> list = bookDao.selectBookList();
 		model.addAttribute("list", list);
 		return "book/bookList";
@@ -30,21 +36,33 @@ public class BookController {
 	}
 	// 입력요청
 	@RequestMapping(value="/book/bookInsert", method = RequestMethod.POST)
-	public String bookInsert(Book book) {
+	public String bookInsert(Book book,HttpSession session) {
+		// 세션에 로그인 값을 확인하고 로그인 정보가 없으면 리다이렉트
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect://login1team/login";
+		}
 		logger.info("입력요청확인 :{}", book);
 		bookDao.insertBook(book);
 		return "redirect:/book/bookList";	// 북 입력 후 "/book/bookList"로 redirect 요청
 	}
 	// 입력페이지요청
 	@RequestMapping(value="/book/bookInsert", method = RequestMethod.GET)
-	public String bookInsert() {
+	public String bookInsert(HttpSession session) {
+		// 세션에 로그인 값을 확인하고 로그인 정보가 없으면 리다이렉트
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect://login1team/login";
+		}
 		logger.debug("입력페이지 요청확인");
 		return "/book/bookInsert";
 	}
 	
 	// 삭제요청
 	@RequestMapping(value="/book/bookDelete", method = RequestMethod.GET)
-	public String bookDelete(@RequestParam(value="bookId", required=true) int bookId) {
+	public String bookDelete(@RequestParam(value="bookId", required=true) int bookId,HttpSession session) {
+		// 세션에 로그인 값을 확인하고 로그인 정보가 없으면 리다이렉트
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect://login1team/login";
+		}
 		logger.debug("삭제 요청확인");
 		bookDao.deleteBook(bookId);
 		return "redirect:/book/bookList";
@@ -52,7 +70,11 @@ public class BookController {
 	
 	// 수정페이지 요청, 수정할 한 권조회
 	@RequestMapping(value="/book/bookUpdate", method = RequestMethod.GET)
-	public String bookOneSelect(Model model,@RequestParam(value="bookId", required=true) int bookId ) {
+	public String bookOneSelect(Model model,@RequestParam(value="bookId", required=true) int bookId,HttpSession session ) {
+		// 세션에 로그인 값을 확인하고 로그인 정보가 없으면 리다이렉트
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect://login1team/login";
+		}
 		logger.debug("수정페이지 요청확인");
 		Book book = bookDao.selectOneBook(bookId);
 		model.addAttribute("book", book);
@@ -61,7 +83,11 @@ public class BookController {
 	
 	// 수정요청
 	@RequestMapping(value="/book/bookUpdate", method = RequestMethod.POST)
-	public String bookUpdate(Book book) {
+	public String bookUpdate(Book book,HttpSession session) {
+		// 세션에 로그인 값을 확인하고 로그인 정보가 없으면 리다이렉트
+		if(session.getAttribute("loginMember")==null) {
+			return "redirect://login1team/login";
+		}
 		logger.info("수정요청확인 :{}", book);
 		bookDao.updateBook(book);
 		return "redirect:/book/bookList";
